Added functionality for User Information

+++ src/main/java/com/we3/codingdashboard/repository/CodingDashboardRepository.java
@@ -3,5 +3,5 @@ package com.we3.codingdashboard.repository;
 import com.we3.codingdashboard.model.CodingDashboard;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface CodingDashboardRepository extends ListCrudRepository<CodingDashboard, Integer> {
+public interface CodingDashboardRepository extends ListCrudRepository<CodingDashboard, Long> {
 }
